fastnext_local_minikube/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
alembic==1.11.1
pydantic==2.3.0
celery==5.3.1
redis==4.5.5
python-dotenv==1.0.0